Trait migrations_internals::MigrationConnection

pub trait MigrationConnection: Connection {
    pub fn setup(&self) -> Result<usize, Error>;
}

A trait indicating that a connection could be used to manage migrations

Only custom backend implementations need to think about this trait

Required methods

pub fn setup(&self) -> Result<usize, Error>

Setup the following table:

diesel::table! {
     __diesel_schema_migrations(version) {
         version -> Text,
         /// defaults to `CURRENT_TIMESTAMP`
         run_on -> Timestamp,
     }
}

Loading content...